// frontend/src/types/generated-schema.ts
type Directus_users = string;
// ...
export interface Projects {
  id?: number;
  status?: string;
  sort?: number | null;
  user_created?: Directus_users | null;
  date_created?: string | null;
  user_updated?: Directus_users | null;
  date_updated?: string | null;
  name: string;
  is_featured?: boolean | null;
  description: string;
  role?: string | null;
  project_type?: string | null;
  start_date?: string | null;
  end_date?: string | null;
  project_url?: string | null;
  github_url?: string | null;
  summary?: string | null;
  // Mark this project as currently active/in progress
  current_project?: boolean | null;
  accomplishments?: unknown | null;
  skills_demonstrated?: unknown | null;
  technologies_used?: unknown | null;
  related_positions?: unknown | null;
  related_companies?: unknown | null;
  related_media?: unknown | null;
}